finished the course

+++ Jose/Token_Refreshing_Jwt_extended/resources/item.py
@@ -1,11 +1,28 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import (
+    jwt_required,
+    get_jwt_claims,
+    jwt_optional,
+    get_jwt_identity,
+    fresh_jwt_required,
+)
 from models.item import ItemModel
 
 
 class Items(Resource):
+    @jwt_optional
     def get(self):
-        return {"items": [item.json() for item in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.query.all()]
+        if user_id:
+            return {"items": items}, 200
+        return (
+            {
+                "items": [item.json()["name"] for item in ItemModel.query.all()],
+                "message": "More data available for logged users",
+            },
+            200,
+        )
 
 
 class Item(Resource):
@@ -18,7 +35,7 @@ class Item(Resource):
         "store_id", help="Every item needs a store id", required=True, type=int
     )
 
-    @jwt_required()
+    @jwt_required
     def get(self, name: str) -> dict:
         item = ItemModel.find_by_name(name)
         if item:
@@ -26,6 +43,7 @@ class Item(Resource):
 
         return {"message": "Item not found"}, 400
 
+    @fresh_jwt_required
     def post(self, name: str) -> dict:
         # Check for duplicate <- error first approach
         if ItemModel.find_by_name(name):
@@ -42,7 +60,13 @@ class Item(Resource):
 
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name: str) -> dict:
+        # jwt makes sure that only logged users will get there
+        claims = get_jwt_claims()
+        if not claims["is_admin"]:
+            return {"message": "Admin privileges required"}
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
